refactor(hogql): relocate property constants and enums to pure modules
Seam 1 / PR 1 of the HogQL de-Django effort (follows #71568). Move the pure
values the HogQL engine imports from Django model modules into Django-free
modules, re-exported from their originals for existing callers:

- posthog/clickhouse/events_json.py <- the events-JSON schema cluster from
  posthog/models/event/sql.py: table-name constants, JSON subcolumn type
  declarations and helpers, index DDL, and EVENTS_JSON_INDEXED_PROPERTY_NAMES.
- posthog/week_start_day.py <- WeekStartDay (posthog/models/team/team.py).
- posthog/property_columns.py <- PropertyName, TableColumn,
  TableWithProperties (posthog/models/property/property.py).
- products/event_definitions/backend/property_type.py <- PropertyType
  (products/event_definitions/backend/models/property_definition.py).

The printer, property planner, and property transforms now import these from
the pure modules, which also removes the Seam 0 lazy-import workarounds for
WeekStartDay and DISTRIBUTED_EVENTS_JSON_TABLE. The printer package itself
stays Django-coupled via the property transforms' materialized-columns
imports; those move behind the property-metadata bundle in the next PRs.

Choice enums relocate verbatim (django.db.models.*Choices subclasses define
without settings or the app registry), so field choices are unchanged and
makemigrations detects no changes.

Generated-By: PostHog Code
Task-Id: e47d9107-4fc1-40d6-8ec6-6cd1be8ea6b0
