+++ posthog/hogql/printer/clickhouse.py
@@ -34,9 +34,10 @@ from posthog.hogql.restricted_properties import restricted_property_keys_for_tab
 from posthog.hogql.type_system import parse_sql_runtime_type
 from posthog.hogql.visitor import GetFieldsTraverser, clone_expr
 
+from posthog.clickhouse.events_json import EVENTS_PROPERTIES_JSON_SUBCOLUMNS, PERSON_PROPERTIES_JSON_SUBCOLUMNS
 from posthog.exchange_rate_constants import EXCHANGE_RATE_DECIMAL_PRECISION, EXCHANGE_RATE_DICTIONARY_NAME
-from posthog.models.event.sql import EVENTS_PROPERTIES_JSON_SUBCOLUMNS, PERSON_PROPERTIES_JSON_SUBCOLUMNS
 from posthog.uuidt import UUIDT
+from posthog.week_start_day import WeekStartDay
 
 
 def _table_filter_type(table_type: ast.TableOrSelectType) -> ast.TableOrSelectType:
@@ -277,8 +278,6 @@ class ClickHousePrinter(BasePrinter):
                 args = [*args[:-1], "6", *args[-1:]]
 
         if node.name == "toStartOfWeek" and len(node.args) == 1:
-            from posthog.models.team.team import WeekStartDay  # noqa: PLC0415
-
             # If week mode hasn't been specified, use the project's default.
             # For Monday-based weeks mode 3 is used (which is ISO 8601), for Sunday-based mode 0 (CH default)
             args.insert(1, WeekStartDay(self._get_week_start_day()).clickhouse_mode)
