+++ products/event_definitions/backend/property_type.py
@@ -1,0 +1,17 @@
+"""Property-type enum, Django-free so the HogQL engine can import it without booting Django.
+
+`django.db.models.TextChoices` is plain enum machinery — defining a subclass touches neither
+settings nor the app registry, so this module imports cleanly in a bare interpreter.
+products.event_definitions.backend.models.property_definition re-exports `PropertyType` for
+existing callers.
+"""
+
+from django.db import models
+
+
+class PropertyType(models.TextChoices):
+    Datetime = "DateTime", "DateTime"
+    String = "String", "String"
+    Numeric = "Numeric", "Numeric"
+    Boolean = "Boolean", "Boolean"
+    Duration = "Duration", "Duration"
